move around some files
