+++ src/simulation/monsters/bosses/wildy/index.ts
@@ -1,0 +1,19 @@
+import Callisto from './Callisto';
+import ChaosElemental from './ChaosElemental';
+import CorporealBeast from './CorporealBeast';
+import CrazyArchaeologist from './CrazyArchaeologist';
+import KingBlackDragon from './KingBlackDragon';
+import Scorpia from './Scorpia';
+import Venenatis from './Venenatis';
+import Vetion from './Vetion';
+
+export const allWildyBosses = {
+	Callisto,
+	ChaosElemental,
+	CrazyArchaeologist,
+	KingBlackDragon,
+	Scorpia,
+	Venenatis,
+	Vetion,
+	CorporealBeast
+};
