+++ src/simulation/monsters/index.ts
@@ -1,57 +1,13 @@
 import Collection from '../../structures/Collection';
 import Monster from '../../structures/Monster';
-import Cerberus from './bosses/Cerberus';
-import CorporealBeast from './bosses/CorporealBeast';
-import DagannothPrime from './bosses/DagannothPrime';
-import DagannothRex from './bosses/DagannothRex';
-import DagannothSupreme from './bosses/DagannothSupreme';
-import GiantMole from './bosses/GiantMole';
-import KalphiteQueen from './bosses/KalphiteQueen';
-import Obor from './bosses/Obor';
-import Sarachnis from './bosses/Sarachnis';
-import Vorkath from './bosses/Vorkath';
-import Zulrah from './bosses/Zulrah';
-import CommanderZilyana from './gwd/CommanderZilyana';
-import GeneralGraardor from './gwd/GeneralGraardor';
-import Kreearra from './gwd/Kreearra';
-import KrilTsutsaroth from './gwd/KrilTsutsaroth';
-import LowMonsters from './low';
-import SpecialMonsters from './special';
-import Callisto from './wildy/Callisto';
-import ChaosElemental from './wildy/ChaosElemental';
-import ChaosFanatic from './wildy/ChaosFanatic';
-import CrazyArchaeologist from './wildy/CrazyArchaeologist';
-import KingBlackDragon from './wildy/KingBlackDragon';
-import Scorpia from './wildy/Scorpia';
-import Venenatis from './wildy/Venenatis';
-import Vetion from './wildy/Vetion';
+import { allBosses } from './bosses';
+import { allLowMonsters } from './low';
+import { specialBosses } from './special';
 
 const monstersObject = {
-	DagannothPrime,
-	DagannothRex,
-	DagannothSupreme,
-	Cerberus,
-	GiantMole,
-	Vorkath,
-	Zulrah,
-	GeneralGraardor,
-	CommanderZilyana,
-	Kreearra,
-	KrilTsutsaroth,
-	Callisto,
-	ChaosElemental,
-	ChaosFanatic,
-	KingBlackDragon,
-	Scorpia,
-	Venenatis,
-	Vetion,
-	CrazyArchaeologist,
-	CorporealBeast,
-	KalphiteQueen,
-	Obor,
-	Sarachnis,
-	...LowMonsters,
-	...SpecialMonsters
+	...allBosses,
+	...allLowMonsters,
+	...specialBosses
 };
 
 const allMonsters: [number, Monster][] = Object.values(monstersObject).map((monster) => [
