+++ src/simulation/monsters/special/index.ts
@@ -2,4 +2,4 @@ import Barrows from './Barrows';
 import Hespori from './Hespori';
 import TzTokJad from './TzTokJad';
 
-export default { Barrows, TzTokJad, Hespori };
+export const specialBosses = { Barrows, TzTokJad, Hespori };
